04_pubsub/requirements.txt
Faker==27.0.0
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-pubsub==2.23.0
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
python-dotenv==1.0.1

